- removed circle loader view - added circle loader to PMSuperbutton layer
